Begin major css refactoring. Create styles folder.
