+++ src/components/CategoryFilter/CategoryFilter.css
@@ -1,4 +1,4 @@
-.categories {
+.categories-wrapper {
   padding-bottom: 1rem;
   overflow-x: auto;
   display: block;
