+++ src/components/Header/Header.css
@@ -1,6 +1,6 @@
 .header {
   margin: 0;
-  padding-top: 0;
+  padding-top: 0.5rem;
   position: sticky;
   top: 0;
   background-color: var(--cream);
@@ -102,23 +102,6 @@
   display: flex;
   align-items: center;
   gap: 0.5rem;
-}
-
-.icon-btn {
-  padding: 0.5rem;
-  color: var(--brown);
-  background: none;
-  border: none;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  transition: color 0.2s;
-  size: 40;
-}
-
-.icon-btn:hover {
-  background: var(--cream_opaque);
-  border-radius: 50%;
-  display: flex;
 }
 
 /* Responsive Design */
