+++ src/components/Recipe/Recipe.css
@@ -1,9 +1,3 @@
-.recipe-container {
-  text-align: left;
-  padding: 0.5rem;
-  padding-top: 0;
-}
-
 .recipe-container h1 {
   margin-top: 0;
   margin-bottom: 0;
@@ -19,25 +13,6 @@
 .recipe-container ol {
   margin-top: 0.5rem;
   margin-bottom: 0.5rem;
-}
-
-.recipe-heading-container {
-  color: var(--dark_red);
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-weight: 600;
-  font-family: forta;
-}
-
-.recipe-subheading {
-  margin-top: 1rem;
-  margin-bottom: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 0.5rem;
 }
 
 .servings {
@@ -64,25 +39,7 @@
   gap: 0.5rem;
 }
 
-.edit-btn,
-.cart-btn {
-  color: var(--dark_red);
-  cursor: pointer;
-  background: none;
-  border: none;
-  transition: color 0.2s, transform 0.2s;
-  padding: 0;
-  display: flex;
-}
 
-.cart-btn:hover,
-.edit-btn:hover {
-  transform: translateY(-1px);
-}
-
-.source-link {
-  color: var(--dark_red);
-}
 
 .cart-counter {
   position: absolute;
