--- src/index.css
@@ -1,101 +1,0 @@
-@import "../public/fonts/fonts.css";
-
-:root {
-  /* Color scheme and background */
-  --cream: #f2e7d2;
-  --cream_opaque: #ffffff99;
-  --light_brown: #523f19bf;
-  --brown: #55443c;
-  --dark_brown: #342a24;
-  --red: #933c1f;
-  --dark_red: #7e341b;
-  --error_red: #ef4444;
-  --error_shadow: #ef444411;
-  --pink: #fee2e2;
-  --dark_pink: #c58293a4;
-  --green: #2a905d;
-  --shadow_black: #0000000d;
-
-  background-color: var(--cream);
-  color-scheme: light dark;
-
-  /* Typography */
-  /* Old font - font-family: "Noto Sans", sans-serif; */
-  font-family: "Gabarito", sans-serif;
-  --heading-font-family: "Forta", sans-serif;
-  font-style: normal;
-  font-weight: 400;
-  font-optical-sizing: auto;
-  font-variation-settings: "wdth" 100;
-  line-height: 1.5;
-  color: var(--dark_brown);
-
-  /* Z-indexes */
-  --z-content: 1;
-  --z-dropdown: 10;
-  --z-header: 100;
-  --z-modal: 1000;
-
-  /* Font rendering */
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-html,
-body {
-  margin: 0;
-  padding: 0;
-}
-
-button {
-  font-size: 1rem;
-  padding: 0.5rem;
-  border-radius: 0.5rem;
-  background: var(--cream_opaque);
-  color: var(--dark_brown);
-  font-family: inherit;
-  border: 0.0625rem solid var(--brown);
-  font-size: 1rem;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-button .auth-btn {
-  min-width: 3rem;
-}
-
-button .shadow-btn {
-  box-shadow: 0 0.125rem 0.5rem var(--shadow_black);
-}
-
-input {
-  padding: 0.5rem;
-  border-radius: 0.5rem;
-  background: var(--cream_opaque);
-  color: var(--dark_brown);
-  font-family: inherit;
-  box-shadow: 0 0.125rem 0.5rem var(--shadow_black);
-  border: 0.0625rem solid var(--brown);
-  font-size: 1rem;
-}
-
-.subheading {
-  padding: 0.5rem 1rem;
-  color: var(--brown);
-  background: none;
-  border: none;
-  cursor: pointer;
-  transition: color 0.2s;
-  font-size: 1rem;
-  font-family: "Forta", sans-serif;
-}
-
-.subheading :hover {
-  color: var(--red);
-}
-
-.subheading.selected {
-  color: var(--red);
-}
