+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 // import { SpeedInsights } from "@vercel/speed-insights/next";
 
-import "./index.css";
+import "../src/styles/index.css";
 import App from "./App.jsx";
 import "./i18n";
