+++ src/styles/App.css
@@ -1,3 +1,5 @@
+@import "./components/buttons.css";
+
 #root {
   margin: 0;
   padding: 0;
