+++ src/styles/globals/variables.css
@@ -1,0 +1,29 @@
+:root {
+  --cream: #f2e7d2;
+  --cream_opaque: #ffffff99;
+  --light_brown: #523f19bf;
+  --brown: #55443c;
+  --dark_brown: #342a24;
+  --red: #933c1f;
+  --dark_red: #7e341b;
+  --error_red: #ef4444;
+  --error_shadow: #ef444411;
+  --pink: #fee2e2;
+  --dark_pink: #c58293a4;
+  --green: #2a905d;
+  --shadow_black: #0000000d;
+
+  --font-size-small: 0.8rem;
+  --font-size-base: 1rem;
+  --font-size-lg: 2rem;
+
+  --z-content: 1;
+  --z-dropdown: 10;
+  --z-header: 100;
+  --z-modal: 1000;
+
+  /* --spacing-xs: 0.25rem;
+  --spacing-sm: 0.5rem;
+  --spacing-md: 1rem;
+  --border-radius: 0.5rem; */
+}
